+++ aggregate.h
@@ -5,7 +5,7 @@ struct s_aggregate {
 	float temperature;
 	float humidity;
 	float rainfall;
-	long pressure;
+	float pressure;
 	float wind_speed;
 	long wind_direction;
 	float light;
